% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgrid.r
\name{achafronteira}
\alias{achafronteira}
\title{Identifica Fronteira Da Região De Convergência}
\usage{
achafronteira(mat)
}
\arguments{
\item{mat}{matriz de verdadeiros e falsos representando a região de convergência}
}
\value{
data.table contendo os númeroes de divisões dos pontos de fronteira
}
\description{
\code{achafronteira} identifica os pontos mais externos da região de convergência
}
\details{
Esta é uma função auxiliar que não deve ser utilizada senão internamente pelo código. É
implementada no contexto de otimização do número de segmentações para geração final do grid.
}
