% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgrid.r
\name{achapersistencia}
\alias{achapersistencia}
\title{Identifica Região De Convergência}
\usage{
achapersistencia(mat)
}
\arguments{
\item{mat}{matriz de verdadeiros e falsos resultante da comparação da matriz de razão de erros
com um parâmetro de razão mínima. Colunas representam de segmentações de vazão turbinada e
linhas de queda líquida}
}
\value{
matriz booleana; TRUE na região de convergência e FALSE do contrário
}
\description{
\code{achapersistencia} identifica dentro de um array de razão de erros a região onde já ocorre
convergência
}
\details{
Esta é uma função auxiliar que não deve ser utilizada senão internamente pelo código. É
implementada no contexto de otimização do número de segmentações para geração final do grid.
}
