% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{agregasemana}
\alias{agregasemana}
\alias{agregasemana.data.table}
\alias{agregasemana.character}
\title{Agregação Semanal}
\usage{
agregasemana(dat, min.horas = 0.9)

\method{agregasemana}{data.table}(dat, min.horas = 0.9)

\method{agregasemana}{character}(dat, min.horas = 0.9)
}
\arguments{
\item{dat}{caminho de arquivo ou \code{data.table} contendo o dado a ser agregado. Ver Detalhes}

\item{min.horas}{percentual indicando o mínimo de horas para considerar a semana valida}
}
\value{
\code{data.table} com médias semanais cotendo as variáveis
\describe{
\item{data}{Data e hora inicial da semana}
\item{nmont}{Nível de montante}
\item{quedal}{Queda líquida media}
\item{perda}{Perda média}
\item{vazao}{Vazão defluente tota}
\item{prod}{Produtibilidade média}
\item{nmaq}{NA, não faz sentido em média semanal}
\item{patamar}{NA, não faz sentido em média semanal}
\item{energia}{NA, não faz sentido em média semanal}
}
}
\description{
Agrega dado horário extraído da planilha em semanas, criticando por um número mínimo de registros
}
\details{
O argumento \code{dat} pode ser tanto um data.table do dado já lido e retornado por
\code{leplanilha} ou uma string contendo o caminho do dado. Neste segundo caso, há duas
possibilidades:
\enumerate{
\item Se \code{dat} é um caminho para arquivo com extensão ".RDS", será lido e agregado
\item Se \code{dat} é um caminho de planilha padrão, com extensão ".xlsm", primeiro será chamada
\cdde{link{leplanilha}} neste arquivo e então é feita a agregação
}
}
