% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgrid.r
\name{extraigrid}
\alias{extraigrid}
\title{Extrator De Grades}
\usage{
extraigrid(fit, dim, ...)
}
\arguments{
\item{fit}{objeto \code{gam(perda)|(prod)} gerado pelas funções de \code{fit} ou \code{opt}
apropriadas}

\item{dim}{dimensão da grade -- vetor de uma posição para perda e duas para produtibilidade
(divisoes de queda liquida, divisoes de vazao)}
}
\value{
object \code{grid(perda)|(prod)}, dependendo da classe do argumento \code{fit}. Em ambos
os casos será uma lista com a grade na primeira posição e o modelo original na segunda
}
\description{
Extrai grades de perda e produtibilidade a parti dos ajustes realizados
}
