% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgam.r
\name{extrapperda}
\alias{extrapperda}
\alias{extrapperda_inf}
\alias{extrapperda_sup}
\title{Extrapolações Para Ajuste De Perdas}
\usage{
extrapperda_inf(dat, mod, tipo, quant_inf)

extrapperda_sup(dat, mod, tipo, quant_sup)
}
\arguments{
\item{dat}{dado ajustado}

\item{mod}{modelo continuo estimado}

\item{tipo}{tipo de extrapolacao. Um entre 0, 1 ou 2}

\item{quant_inf, quant_sup}{o quantil para extracao do dado ajustado na extrapolacao}
}
\value{
lista de dois elementos: o vetor de coeficientes ajustado e valor do corte no domínio
}
\description{
Funções auxiliares para \code{fitgam_perda}
}
