% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgam.r
\name{fitgam_perda}
\alias{fitgam_perda}
\title{Ajuste De Curva Para Perdas}
\usage{
fitgam_perda(
  dat,
  ns.vazao = 10,
  ts.vazao = "tp",
  extrap = c(1, 1),
  quantil = c(0.05, 0.95)
)
}
\arguments{
\item{dat}{\code{data.table} de dados para ajuste. Ver Detalhes}

\item{ns.vazao}{numero de nós no GAM ajustado. Padrao 10}

\item{ts.vazao}{tipo de spline utilizada para vazao -- um de \code{c("tp", "cr")}; veja
\code{link[mgcv]{gam}} e Detalhes. Padrao "tp"}

\item{extrap}{vetor de duas posições indicando o tipo de extrapolação em cada região. Ver
Detalhes}

\item{quantil}{quantis para uso na extrapolação, caso seja tipo 2. Ver Detalhes}
}
\value{
objeto \code{gamperda} contendo GAM e extrapolações estimadas
}
\description{
Estima GAM para perdas, complementando regiões inferior e superior com extrapolação
}
\details{
\code{fitgam_perda} espera que alguns atributos especiais estejam presetnes no argumento
\code{dat}, dentre eles: codigo, nome e vazao turbinada efetiva total da usina. Dentre estes,
apenas o último é relevante para a estimação de curva, sendo os outros dois necessários apenas
para escrita de resultados e plots. Desta forma, para garantir o funcionamento apropriado da
função o objeto passado em \code{dat} deve ter sido lido e (possivelmente) agregado semanalmente
pelas funções fornecidas no pacote.
}
