% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgam.r
\name{fitgam_prod}
\alias{fitgam_prod}
\title{Ajuste De Curva Para Perdas}
\usage{
fitgam_prod(
  dat,
  ns.quedal = 10,
  ns.vazao = 10,
  ts.quedal = "ps",
  ts.vazao = "ps",
  bordas = TRUE,
  modo = c("tensor", "multivar", "simples")
)
}
\description{
Estima GAM para perdas, complementando regiões inferior e superior com extrapolação
}
