% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{fitted.gamperda}
\alias{fitted.gamperda}
\title{Valores Ajustados De Objetos \code{gamperda}}
\usage{
\method{fitted}{gamperda}(object, ...)
}
\arguments{
\item{object}{objeto \code{gamperda} retornado por \code{\link{fitgam_perda}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de valores ajustados pelo modelo estimado
}
\description{
Método \code{fitted} para objetos da classe \code{gamperda}
}
\seealso{
Other metodos gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{optgam_perda}()},
\code{\link{predict.gamperda}()},
\code{\link{residuals.gamperda}()}
}
\concept{metodos gamperda}
