% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamprod.r
\name{fitted.gamprod}
\alias{fitted.gamprod}
\title{Valores Ajustados De Objetos \code{gamprod}}
\usage{
\method{fitted}{gamprod}(object, ...)
}
\arguments{
\item{object}{objeto \code{gamprod} retornado por \code{\link{fitgam_prod}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de valores ajustados pelo modelo estimado
}
\description{
Método \code{fitted} para objetos da classe \code{gamprod}
}
\seealso{
Other metodos gamprod: 
\code{\link{fitgam_prod}()},
\code{\link{optgam_prod}()},
\code{\link{predict.gamprod}()},
\code{\link{residuals.gamprod}()}
}
\concept{metodos gamprod}
