% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{fitted.gridperda}
\alias{fitted.gridperda}
\title{Valores Ajustados De Objetos \code{gridperda}}
\usage{
\method{fitted}{gridperda}(object, ...)
}
\arguments{
\item{object}{objeto \code{gridperda} retornado por \code{\link{extraigrid.gamperda}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de valores ajustados pelo modelo estimado
}
\description{
Método \code{fitted} para objetos da classe \code{gridperda}
}
\seealso{
Other metodos gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{optgrid}()},
\code{\link{predict.gridperda}()},
\code{\link{residuals.gridperda}()}
}
\concept{metodos gridperda}
