% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridprod.r
\name{fitted.gridprod}
\alias{fitted.gridprod}
\title{Valores Ajustados De Objetos \code{gridprod}}
\usage{
\method{fitted}{gridprod}(object, ...)
}
\arguments{
\item{object}{objeto \code{gridprod} retornado por \code{\link{extraigrid.gamprod}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de valores ajustados pelo modelo estimado
}
\description{
Método \code{fitted} para objetos da classe \code{gridprod}
}
\seealso{
Other metodos gridprod: 
\code{\link{predict.gridprod}()},
\code{\link{residuals.gridprod}()}
}
\concept{metodos gridprod}
