% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r, R/visualizacao.r
\name{gamperda}
\alias{gamperda}
\alias{new_gamperda}
\alias{predict.gamperda}
\alias{fitted.gamperda}
\alias{residuals.gamperda}
\alias{plot.gamperda}
\alias{lines.gamperda}
\title{Objeto \code{gamperda}}
\usage{
new_gamperda(dat, mod, coefI, coefS, corteI, corteS, fitcall)

\method{predict}{gamperda}(object, newdata, ...)

\method{fitted}{gamperda}(object, ...)

\method{residuals}{gamperda}(object, ...)

\method{plot}{gamperda}(x, legenda = TRUE, ...)

\method{lines}{gamperda}(x, ...)
}
\arguments{
\item{dat}{dado utilizado para ajuste}

\item{mod}{modelo ajustado}

\item{coefI, coefS}{coeficientes ajustados das extrapolações inferior e superior}

\item{corteI, corteS}{valores dos cortes entre extrapolações e modelo aditivo}

\item{fitcall}{chamada de \code{fitgam_perda} original}

\item{object}{objeto \code{gamperda} retornado por \code{\link{fitgam_perda}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}

\item{...}{demais parametros}
}
\value{
para \code{predict.gamperda}, vetor de perdas previstas nas abscissas contidas em
\code{newdata}. Para code{fitted.gamperda}, os valores ajustados e \code{rediduals.gamperda}
os erros
}
\description{
Construtor e métodos da classe \code{gamperda}

Função para contrução da saída de \code{fitgam_perda}, não deve ser chamada diretamente
}
