% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamprod.r
\name{gamprod}
\alias{gamprod}
\alias{new_gamprod}
\title{Objeto \code{gamprod}}
\usage{
new_gamprod(dat, mod, atributos, args)
}
\description{
Construtor e métodos da classe \code{gamprod}

Função para contrução da saída de \code{fitgam_perda}, não deve ser chamada diretamente
}
