% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{extraigrid.gamperda}
\alias{extraigrid.gamperda}
\alias{new_gridperda}
\alias{predict.gridperda}
\alias{fitted.gridperda}
\alias{residuals.gridperda}
\title{Extrator De Grades de Perda}
\usage{
\method{extraigrid}{gamperda}(fit, dim, ...)

new_gridperda(perda, vazao, fit)

\method{predict}{gridperda}(object, newdata, ...)

\method{fitted}{gridperda}(object, ...)

\method{residuals}{gridperda}(object, ...)
}
\arguments{
\item{fit}{object \code{gamperda} retornado por \code{fitgam_perda} ou \code{optgam_perda}}

\item{dim}{inteiro indicando número de pontos na grade a ser extraída}

\item{...}{demais parametros}

\item{object}{objeto \code{gridperda} retornado por \code{\link{extraigrid.gamperda}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}
}
\value{
objeto da classe \code{gridperda} contendo grade extraída e ajuste original

objeto \code{gridperda} contendo grade extraída e modelo originalmente ajustado ao dado
}
\description{
Amostra \code{dim} pontos igualmente espaçacados no ajuste da curva de perdas

Função para contrução da saída de \code{extraigrid.gamperda}, não deve ser chamada diretamente
}
