% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{INTERPBILIN}
\alias{INTERPBILIN}
\title{Interpolação Bilinear}
\usage{
INTERPBILIN(xData, yData, zData, xPred, yPred)
}
\arguments{
\item{xData, yData, zData}{valores das variáveis explicativas (x e y) e dependente (z), na grade}

\item{xPred, yPred}{valores das variáveis explicativas a interpolar na grade}
}
\value{
vetor de valores interpolados
}
\description{
Realiza interpolação bilinear em uma grade tridimensional de valores
}
