% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{INTERPLIN}
\alias{INTERPLIN}
\title{Interpolação Linear}
\usage{
INTERPLIN(yData, zData, yPred)
}
\arguments{
\item{yData, zData}{valores da variável explicativa e dependente, respectivamente, na grade}

\item{yPred}{valores da variável explicativa a interpolar na grade}
}
\value{
vetor de valores interpolados
}
\description{
Realiza interpolação linear em uma grade bidimensional de valores
}
