% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{leaba}
\alias{leaba}
\title{Leitura De Uma Aba Da Planilha Padrão}
\usage{
leaba(arq, aba, range, nomes, tipos)
}
\arguments{
\item{arq}{caminho completo da planilha}

\item{aba}{nome ou índice da aba a ser lida}

\item{range}{range no formato excel de células para ler, incluindo celulas com nome das colunas}

\item{nomes}{nomes para atribuir às colunas}

\item{tipos}{tipo de dado em cada coluna sendo lida}
}
\value{
\code{data.table} contendo o range na aba especificados
}
\description{
Função interna para leitura das múltiplas abas da planilha padrão
}
