% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{lines.gamperda}
\alias{lines.gamperda}
\title{Plot De Linha Dos Modelos Ajustados Para Perda}
\usage{
\method{lines}{gamperda}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{gamperda}}

\item{...}{demais parâmetros passados à genérica \code{\link[graphics]{lines}}}
}
\value{
plota a linha do ajuste do modelo contido em \code{x}
}
\description{
Wrapper para visualização apenas dos modelos ajustados, sem dados ao fundo
}
\examples{

dat <- agregasemana(dummydata)
mod1 <- fitgam_perda(dat)
mod2 <- fitgam_perda(dat, 15)

\dontrun{
plot(mod1)
lines(mod2, col = 3, lwd = 2. lty = 2)
}

}
\seealso{
Other plots gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{optgam_perda}()},
\code{\link{plot.gamperda}()}
}
\concept{plots gamperda}
