% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{new_gamperda}
\alias{new_gamperda}
\title{Construtor Interno}
\usage{
new_gamperda(dat, mod, coefI, coefS, corteI, corteS, fitcall)
}
\arguments{
\item{dat}{dado utilizado para ajuste}

\item{mod}{modelo ajustado}

\item{coefI, coefS}{coeficientes ajustados das extrapolações inferior e superior}

\item{corteI, corteS}{valores dos cortes entre extrapolações e modelo aditivo}

\item{fitcall}{chamada de \code{fitgam_perda} original}
}
\value{
objeto \code{gamperda} com modelo ajustado e dado original
}
\description{
Função para contrução da saída de \code{fitgam_perda}, não deve ser chamada diretamente
}
