% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{new_gridperda}
\alias{new_gridperda}
\title{Construtor Interno}
\usage{
new_gridperda(perda, vazao, fit)
}
\arguments{
\item{perda}{perda amostrada do modelo}

\item{vazao}{vazoes correspondentes às segmentações da grade}

\item{fit}{modelo ajustado aos dados}
}
\value{
objeto \code{gridperda} com a grade extraída e modelo original
}
\description{
Função para contrução da saída de \code{extraigrid.gamperda}, não deve ser chamada diretamente
}
