% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgam.r
\name{optgam_perda}
\alias{optgam_perda}
\title{Otimização Do Ajuste De Perdas}
\usage{
optgam_perda(
  dat,
  range.vazao = 5:30,
  ts.vazao = "ps",
  extrap = c(2, 2),
  quantil = c(0.05, 0.95)
)
}
\arguments{
\item{dat}{\code{data.table} de dados para ajuste. Ver Detalhes}

\item{range.vazao}{vetor de numero de nós a testar na seleção. Padrao 5:30}

\item{ts.vazao}{tipo de spline utilizada para vazao -- um de \code{c("tp", "cr", "ps")}; veja
\code{link[mgcv]{gam}} e Detalhes. Padrao "ps"}

\item{extrap}{vetor de duas posições indicando o tipo de extrapolação em cada região. Ver
Detalhes. Padrao tipo 2 para ambas as extrapolações}

\item{quantil}{quantis para uso na extrapolação. Ver Detalhes}
}
\value{
objeto \code{gamperda} contendo GAM e extrapolações estimadas
}
\description{
Varre uma faixa de dimensões de base, reportando aquele correspondente ao ajuste de menor BIC
}
\details{
Os parâmetros recebidos por \code{optgam_perda} são exatamente aqueles de
\code{\link{fitgam_perda}}, com uma exceção: o argumento \code{ns.vazao} é substituído por
\code{range.vazao}, um vetor de inteiros indicando as dimensões de base candidatos para seleção.
}
\examples{

dat <- agregasemana(dummydata)

# execucao limitando a faixa de numero de splines a algo mais baixo
optfit <- optgam_perda(dat, range.vazao = 5:10)
plot(optfit)

# dentre os argumentos de optfit esta a lista 'gamargs', contendo a parametrizacao do gam
attr(optfit, "gamargs") # ns.vazao corresponde ao numero de nos utilizados no ajuste otimo

}
\seealso{
ajuste individual para uma dada dimensão de base \code{\link{fitgam_perda}}

Other metodos gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{fitted.gamperda}()},
\code{\link{predict.gamperda}()},
\code{\link{residuals.gamperda}()}

Other plots gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{plot.gamperda}()}
}
\concept{metodos gamperda}
\concept{plots gamperda}
