% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgam.r
\name{optgam_prod}
\alias{optgam_prod}
\title{Otimização Do Ajuste De Produtibilidade}
\usage{
optgam_prod()
}
\description{
Varre uma faixa de dimensões de base, reportando aquele correspondente ao ajuste de menor BIC
}
