% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perdaprodvar.r
\docType{package}
\name{perdaprodvar}
\alias{perdaprodvar}
\title{\code{perdaprodvar}}
\description{
Modelagem De Dados De Perda E Produtibillidade E Extração De Grades
}
\details{
Este pacote contém o conjunto de funções necessárias para aplicação e análise da metodologia para
modelagem de perda e produtibillidade variáveis através de modelos aditivos. Além de funções
dedicadas à extração dos dados e cálculo das médias semanais, são fornecidas funções para ajuste
dos modelos, otimização dos hiperparâmetros dos mesmos, extração de grades e otimização do número
de segmentações necessárias, segundo critérios acordados no GT Produtibillidade Hidrelétrica da
CPAMP. Finalmente, funções para visualização e diagnósticos são inclusas.
}
