% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.gamperda}
\alias{plot.gamperda}
\title{Plot Completo De Objetos \code{gamperda}}
\usage{
\method{plot}{gamperda}(x, ..., plot = TRUE)
}
\arguments{
\item{x}{objeto da classe \code{gamperda}}

\item{...}{outros objetos \code{gamperda} para adicionar ao plot}

\item{plot}{booleano indicando se o plot deve ser gerado ou apenas retornado invisivelmente}
}
\value{
plot dos dados originais e ajuste(s) realizado(s)
}
\description{
Wrapper para visualização dos modelos contínuos de perda junto aos dados ajustados
}
\examples{

dat <- agregasemana(dummydata)
mod1 <- fitgam_perda(dat)
mod2 <- fitgam_perda(dat, 5)
mod3 <- fitgam_perda(dat, 7)

\dontrun{
plot(mod1)
plot(mod1, mod2, mod3)
}

}
\seealso{
Other plots gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{optgam_perda}()}
}
\concept{plots gamperda}
