% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.gamprod}
\alias{plot.gamprod}
\title{Plot Completo De Objetos \code{gamprod}}
\usage{
\method{plot}{gamprod}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{gamprod}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
plota dados originais e ajuste realizado. Como o plot é 3d através do pacote
\code{plotly}, é retornado um objeto desta classe contendo o plot para posteriores
modificações
}
\description{
Wrapper para visualização dos modelos contínuos de produtibilidade junto aos dados ajustados
}
\examples{

dat <- agregasemana(dummydata)
mod <- fitgam_prod(dat)

\dontrun{
plot(mod)
}

}
\seealso{
Other plots gamprod: 
\code{\link{fitgam_prod}()}
}
\concept{plots gamprod}
