% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.gridperda}
\alias{plot.gridperda}
\title{Plot Completo De Objetos \code{gridperda}}
\usage{
\method{plot}{gridperda}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{objeto da classe \code{gridperda}}

\item{plot}{booleano indicando se o plot deve ser gerado ou apenas retornado invisivelmente}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
plota dados originais, ajuste realizado e grade de perdas extraída
}
\description{
Wrapper para visualização dos modelos contínuos de perda, dados ajustados e grade extraída
}
\examples{

dat <- agregasemana(dummydata)
mod <- fitgam_perda(dat)
grd <- extraigrid(mod, 20)

\dontrun{
plot(grd)
}

}
\seealso{
Other plots gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{optgrid}()}
}
\concept{plots gridperda}
