% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.gridprod}
\alias{plot.gridprod}
\title{Plot De Objetos \code{gridprod}}
\usage{
\method{plot}{gridprod}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{objeto da classe \code{gridprod}}

\item{plot}{booleano indicando se o plot deve ser gerado ou apenas retornado invisivelmente}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
plota dados originais, ajuste realizado e grade de prods extraída
}
\description{
Wrapper para visualização dos dados ajustados e grade de produtibilidade extraída
}
\examples{

dat <- agregasemana(dummydata)
mod <- fitgam_prod(dat)
grd <- extraigrid(mod, 20)

\dontrun{
plot(grd)
}

}
\concept{plots gridprod}
