% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.varreduraprod}
\alias{plot.varreduraprod}
\title{Plot Da Varredura De Número De Divisões Para Produtibilidade}
\usage{
\method{plot}{varreduraprod}(x, plot = TRUE, ...)
}
\arguments{
\item{x}{objeto da classe \code{gridprod}}

\item{plot}{booleano indicando se o plot deve ser gerado ou apenas retornado invisivelmente}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
plota dados originais, ajuste realizado e grade de produtibilidade extraída
}
\description{
Wrapper para visualização do resultado da varredura realizada na otimização do tamanho de grade
}
\examples{

dat <- agregasemana(dummydata)
mod <- fitgam_prod(dat)
grd <- optgrid(mod, full.output = TRUE)

\dontrun{
plot(grd[[2]])
}

}
