% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{predict.gamperda}
\alias{predict.gamperda}
\title{Previsão Com Objetos \code{gamperda}}
\usage{
\method{predict}{gamperda}(object, newdata, ...)
}
\arguments{
\item{object}{objeto \code{gamperda} retornado por \code{\link{fitgam_perda}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de perdas previstas nas abscissas contidas em \code{newdata}
}
\description{
Método \code{predict} para objetos da classe \code{gamperda}
}
\seealso{
Other metodos gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{fitted.gamperda}()},
\code{\link{residuals.gamperda}()}
}
\concept{metodos gamperda}
