% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamprod.r
\name{predict.gamprod}
\alias{predict.gamprod}
\title{Previsão Com Objetos \code{gamprod}}
\usage{
\method{predict}{gamprod}(object, newdata, ...)
}
\arguments{
\item{object}{objeto \code{gamprod} retornado por \code{\link{fitgam_prod}}}

\item{newdata}{data.frame ou data.table contendo variáveis explicativas para previsão}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de produtibilidades previstas nas abscissas contidas em \code{newdata}
}
\description{
Método \code{predict} para objetos da classe \code{gamprod}
}
\seealso{
Other metodos gamprod: 
\code{\link{fitgam_prod}()},
\code{\link{fitted.gamprod}()},
\code{\link{optgam_prod}()},
\code{\link{residuals.gamprod}()}
}
\concept{metodos gamprod}
