% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{predict.gridperda}
\alias{predict.gridperda}
\title{Previsão Com Objetos \code{gridperda}}
\usage{
\method{predict}{gridperda}(object, newdata, ...)
}
\arguments{
\item{object}{objeto \code{gridperda} retornado por \code{\link{extraigrid.gamperda}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de perdas previstas nas abscissas contidas em \code{newdata}
}
\description{
Método \code{predict} para objetos da classe \code{gridperda}
}
\seealso{
Other metodos gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{fitted.gridperda}()},
\code{\link{optgrid}()},
\code{\link{residuals.gridperda}()}
}
\concept{metodos gridperda}
