% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridprod.r
\name{predict.gridprod}
\alias{predict.gridprod}
\title{Previsão Com Objetos \code{gridprod}}
\usage{
\method{predict}{gridprod}(object, newdata, ...)
}
\arguments{
\item{object}{objeto \code{gridprod} retornado por \code{\link{extraigrid.gamprod}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de prods previstas nas abscissas contidas em \code{newdata}
}
\description{
Método \code{predict} para objetos da classe \code{gridprod}
}
\seealso{
Other metodos gridprod: 
\code{\link{fitted.gridprod}()},
\code{\link{residuals.gridprod}()}
}
\concept{metodos gridprod}
