% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamprod.r
\name{print.gamprod}
\alias{print.gamprod}
\title{Print De Objetos \code{gamprod}}
\usage{
\method{print}{gamprod}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{gamprod}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
print do modelo no console, sem retornar nenhum valor
}
\description{
Print De Objetos \code{gamprod}
}
