% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{residuals.gamperda}
\alias{residuals.gamperda}
\title{Resíduos De Objetos \code{gamperda}}
\usage{
\method{residuals}{gamperda}(object, ...)
}
\arguments{
\item{object}{objeto \code{gamperda} retornado por \code{\link{fitgam_perda}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de erros de ajuste do modelo estimado
}
\description{
Método \code{residuals} para objetos da classe \code{gamperda}
}
\seealso{
Other metodos gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{fitted.gamperda}()},
\code{\link{optgam_perda}()},
\code{\link{predict.gamperda}()}
}
\concept{metodos gamperda}
