% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamprod.r
\name{residuals.gamprod}
\alias{residuals.gamprod}
\title{Resíduos De Objetos \code{gamprod}}
\usage{
\method{residuals}{gamprod}(object, ...)
}
\arguments{
\item{object}{objeto \code{gamprod} retornado por \code{\link{fitgam_prod}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de erros de ajuste do modelo estimado
}
\description{
Método \code{residuals} para objetos da classe \code{gamprod}
}
\seealso{
Other metodos gamprod: 
\code{\link{fitgam_prod}()},
\code{\link{fitted.gamprod}()},
\code{\link{predict.gamprod}()}
}
\concept{metodos gamprod}
