% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{residuals.gridperda}
\alias{residuals.gridperda}
\title{Resíduos De Objetos \code{gridperda}}
\usage{
\method{residuals}{gridperda}(object, ...)
}
\arguments{
\item{object}{objeto \code{gridperda} retornado por \code{\link{extraigrid.gamperda}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de erros de ajuste do modelo estimado
}
\description{
Método \code{residuals} para objetos da classe \code{gridperda}
}
\seealso{
Other metodos gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{fitted.gridperda}()},
\code{\link{predict.gridperda}()}
}
\concept{metodos gridperda}
