% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridprod.r
\name{residuals.gridprod}
\alias{residuals.gridprod}
\title{Resíduos De Objetos \code{gridprod}}
\usage{
\method{residuals}{gridprod}(object, ...)
}
\arguments{
\item{object}{objeto \code{gridprod} retornado por \code{\link{extraigrid.gamprod}}}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
vetor de erros de ajuste do modelo estimado
}
\description{
Método \code{residuals} para objetos da classe \code{gridprod}
}
\seealso{
Other metodos gridprod: 
\code{\link{fitted.gridprod}()},
\code{\link{predict.gridprod}()}
}
\concept{metodos gridprod}
