% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dummydata}
\alias{dummydata}
\title{Exemplo De Dado Semanal}
\format{
Dado horário de uma usina do SIN reescalonado de modo a manter a anonimidade. Contem as variáveis

\describe{
\item{data}{Data e hora inicial da semana}
\item{nmont}{Nível de montante}
\item{quedal}{Queda líquida media}
\item{perda}{Perda média}
\item{vazao}{Vazão defluente tota}
\item{prod}{Produtibilidade média}
\item{nmaq}{NA, não faz sentido em média semanal}
\item{patamar}{NA, não faz sentido em média semanal}
\item{energia}{NA, não faz sentido em média semanal}
}
}
\usage{
dummydata
}
\description{
Exemplo De Dado Semanal
}
\keyword{datasets}
