% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{dummydata}
\alias{dummydata}
\title{Exemplo De Dado Semanal}
\format{
\code{data.table} com as colunas\preformatted{\\describe\{
    \\item\{datahora\}\{Data e hora do registro\}
    \\item\{patamar\}\{Patamar de carga\}
    \\item\{energia\}\{Energia gerada total da usina\}
    \\item\{prod\}\{Produtibilidade média\}
    \\item\{vazao\}\{Vazão defluente tota\}
    \\item\{nmaq\}\{Número de máquinas em operação\}
    \\item\{nmont\}\{Nível de montante\}
    \\item\{perda\}\{Perda média\}
    \\item\{quedal\}\{Queda líquida média\}
\}
}
}
\usage{
dummydata
}
\description{
Dado horário de uma usina do SIN reescalonado de modo a manter a anonimidade.
}
\keyword{datasets}
