% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitgam.r
\name{fitgam_prod}
\alias{fitgam_prod}
\title{Ajuste De Superfície Para Produtibilidade}
\usage{
fitgam_prod(
  dat,
  ns = c(10, 10),
  ts = c("ps", "ps"),
  dist = gaussian(),
  bordas = TRUE,
  modo = c("tensor", "multivar", "simples"),
  gamctrl = list()
)
}
\arguments{
\item{dat}{\code{data.table} de dados para ajuste. Ver Detalhes}

\item{ns}{vetor dois inteiros, indicando dimensão da base em queda líquida e vazão, nesta ordem}

\item{ts}{vetor de duas strings tipo de spline utilizada para queda líquida e vazão,
respectivamente -- veja \code{\link[mgcv]{smooth.terms}} para mais detalhes sobre as opções}

\item{dist}{objeto da classe \code{family} indicando a distribuição e link a serem usados no
ajuste. Veja \code{\link[stats]{family}} e \code{\link[stats]{glm}} para mais informações.}

\item{bordas}{booleano indicando o uso ou não de bordas; alternativamente, um vetor de inteiros
indicando quais vértices utilizar. Ver Detalhes}

\item{modo}{um de \code{"tensor"}, \code{"multivar"} ou \code{"simples"} indicando o modo de
interação entre funções marginais. Ver Detalhes}

\item{gamctrl}{lista contendo parâmetros de controle para a estimação do modelo aditivo. Ver
\code{\link[mgcv]{gam.control}} (ou \code{\link[scam]{scam.control}} quando usando shape
constrained splines)}
}
\value{
objeto \code{gamprod} contendo modelo aditivo ajustado
}
\description{
Estima GAM para produtibilidades
}
\details{
\code{fitgam_prod} espera que alguns atributos especiais estejam presetnes no argumento
\code{dat}, dentre eles: codigo, nome e vazao turbinada efetiva total da usina. Destes, apenas
o último é relevante para a estimação de curva, sendo os outros dois necessários para escrita
de resultados e plots. Desta forma, para garantir o funcionamento apropriado da função o objeto
passado em \code{dat} deve ter sido lido e (possivelmente) agregado semanalmente pelas funções
fornecidas no pacote.

\code{ts} permite a especificação do tipo de spline utilizada na expansão de base em cada
marginal. A maioria das opções suportadas pelo pacote \code{mgcv} são utilizáveis, com algumas
exceções que não condizem com esta aplicação específica. \code{fitgam_prod} suporta as splines
livres

\itemize{
\item \code{"tp"}: thin-plate regression splines
\item \code{"ts"}: thin-plate regression splines com encolhimento
\item \code{"ds"}: duchon splines
\item \code{"cr"}: cubic regression splines
\item \code{"cs"}: cubic regression splines com encolhimento
\item \code{"ps"}: P-splines
}

Para maiores detalhes a respeito das possibilidades e suas descrições, veja
\code{\link[mgcv]{smooth.terms}}. Por padrão é utilizado  \code{ts = c("ps", "ps")}, o que
corresponde à expansão por P-Splines.

Adicionalmente às splines denominadas livres, \code{fitgam_prod} também suporta o uso de shape
constrained P-splines, especificamente uma reparametrização que garante uma superfície côncava no
domínio de ajuste, via \code{ts = c("cv", "cv")}. Maiores detalhes a respeito desta modelagem
encontram-se em \code{link[scam]{scam}} e \code{\link[scam]{shape.constrained.smooth.terms}}.
Embora existam ainda outras diversas parametrizações de shape constraints, apenas a concavidade
faz sentido para produtibilidade.

Embora a função ofereça a possibilidade de utilizar ou não as bordas da curva colina como apoio
para estimação do modelo, o domínio final é sempre o mesmo:
\itemize{
\item o ajuste de produtibilidade sempre cobre da vazão zero até o máximo entre a vazão histórica
e aquela estipulada no HIDR, ou seja, mesmo que colina tenha bordas em vazões superiores, o
ajuste NÃO vai até este ponto, dado que o DECOMP não consultará vazões tão elevadas
\item o domínio de queda líquida, por outro lado, sempre vai da menor a maior queda observadas na
curva colina, mesmo que a borda não seja contemplada no ajuste. Isto é feito de modo a
garantir que o ajuste possua o domínio completo de quedas operáveis pelas máquinas da usina
}

Ainda no que toca a borda, seu uso pode ser informado de duas formas: um valor lógico ou vetor
numérico. Caso o primeiro seja informado, todos os pontos da borda são utilizados ou não. Caso
se informe um vetor numérico, apenas os pontos listados neste vetor serão utilizados. A
disposição do dado de borda é:
\enumerate{
\item menor queda, menor vazão
\item MAIOR queda, menor vazão
\item menor queda, MAIOR vazão
\item MAIOR queda, MAIOR vazão
}

O argumento \code{modo} permite definir a estrutura geral do modelo aditivo.
\code{modo = "simples"} implica em um modelo de efeitos aditivos sem iteração, isto é, a função
estimada é da forma \eqn{g(x) = f_1(X_1) + f_2(X_2)}, sem interação entre as variáveis.
\code{modo = "tensor"} construirá uma função por produto tensor das marginais, isto é, levando em
consideração a interação entre variáveis. \code{modo = "multivar"} só é aplicável quando
\code{all(ts \%in\% c("tp", "ts"))}, pois corresponde a modelagem com suavizadores
naturalmente multivariados, que é o caso singular das thin plate regression splines.
}
\examples{

dat <- agregasemana(dummydata)
# dimensao de base 8 para queda e 10 para vazao -- thin plate splines para queda e vazao
mod <- fitgam_prod(dat, ns = c(8, 10), ts = c("tp", "tp"))

# valores ajustados, previsao e residuos
res <- residuals(mod)
fitt <- fitted(mod)
predd <- predict(mod, newdata = data.frame(quedal = runif(100, 20.5, 21),
vazao = runif(100, 20, 150)))

# objetos retornados por fitgam_prod possuem um metodo de plot e lines, para facil visualizacao
\dontrun{
plot(mod)
}

}
\seealso{
otimização da dimensão de base \code{\link{optgam_prod}}.

Other metodos gamprod: 
\code{\link{fitted.gamprod}()},
\code{\link{optgam_prod}()},
\code{\link{predict.gamprod}()},
\code{\link{residuals.gamprod}()}

Other plots gamprod: 
\code{\link{optgam_prod}()},
\code{\link{plot.gamprod}()}
}
\concept{metodos gamprod}
\concept{plots gamprod}
