% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{gamperda}
\alias{gamperda}
\alias{new_gamperda}
\alias{predict.gamperda}
\alias{fitted.gamperda}
\alias{residuals.gamperda}
\title{Objeto \code{gamperda}}
\usage{
new_gamperda(dat, mod, coefI, corteI, coefS, corteS, atributos)

\method{predict}{gamperda}(object, newdata, ...)

\method{fitted}{gamperda}(object, ...)

\method{residuals}{gamperda}(object, ...)
}
\arguments{
\item{object}{objeto \code{gamperda} retornado por \code{\link{fitgam_perda}}}

\item{newdata}{data.frame ou data.table contendo variável explicativa para previsão}

\item{...}{demais parametros}
}
\value{
para \code{predict.gamperda}, vetor de perdas previstas nas abscissas contidas em
\code{newdata}. Para code{fitted.gamperda}, os valores ajustados e \code{rediduals.gamperda} os
erros
}
\description{
Construtor e métodos da classe \code{gamperda}

Função para contrução da saída de \code{fitgam_perda}, não deve ser chamada diretamente
}
