% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leitura.r
\name{leplanilha}
\alias{leplanilha}
\title{Interface Com a Planilha Padrão}
\usage{
leplanilha(arq)
}
\arguments{
\item{arq}{caminho completo do arquivo a ser lido}
}
\value{
\code{data.table} em base horária contendo as variáveis
\describe{
\item{datahora}{Data e hora do registro}
\item{nmont}{Nível de montante}
\item{quedal}{Queda líquida media}
\item{perda}{Perda média}
\item{vazao}{Vazão defluente tota}
\item{prod}{Produtibilidade média}
\item{nmaq}{Número de máquinas em operação}
\item{patamar}{Patamar de carga}
\item{energia}{Energia gerada total}
}
}
\description{
Realiza a leitura e tratamento dos dados na planilha padrão do GTDP
}
