% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgam.r
\name{optgam_perda}
\alias{optgam_perda}
\title{Otimização Do Ajuste De Perdas}
\usage{
optgam_perda(
  dat,
  range = 5:30,
  ts = "ps",
  dist = gaussian(),
  extrap = c(2, 2),
  quantil = c(0.05, 0.95)
)
}
\arguments{
\item{dat}{\code{data.table} de dados para ajuste. Ver Detalhes}

\item{range}{vetor de dimensões de base a testar na seleção}

\item{ts}{tipo de spline utilizada para vazao -- um de \code{c("tp", "cr", "ps")}; veja
\code{link[mgcv]{gam}} e Detalhes}

\item{dist}{objeto da classe \code{family} indicando a distribuição e link a serem usados no
ajuste. Veja \code{\link[stats]{family}} e \code{\link[stats]{glm}} para mais informações.}

\item{extrap}{vetor de duas posições indicando o tipo de extrapolação em cada região. Ver
Detalhes}

\item{quantil}{quantis para uso na extrapolação. Ver Detalhes}
}
\value{
objeto \code{gamperda} contendo GAM e extrapolações estimadas
}
\description{
Varre uma faixa de dimensões de base, reportando aquela correspondente ao ajuste de menor BIC
}
\details{
Os parâmetros recebidos por \code{optgam_perda} são exatamente aqueles de
\code{\link{fitgam_perda}}, com uma exceção: o argumento \code{ns} é substituído por
\code{range}, um vetor de inteiros indicando as dimensões de base candidatas para seleção.
Mais detalhes a respeito da estimação e efeito dos argumentos individuais podem ser encontrados
em \code{\link{fitgam_perda}}
}
\examples{

dat <- agregasemana(dummydata)

# execucao limitando a faixa de numero de splines a algo mais baixo
optfit <- optgam_perda(dat, range = 5:10)
\dontrun{
plot(optfit)
}

}
\seealso{
ajuste individual para uma dada dimensão de base \code{\link{fitgam_perda}}

Other metodos gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{fitted.gamperda}()},
\code{\link{predict.gamperda}()},
\code{\link{residuals.gamperda}()}

Other plots gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{plot.gamperda}()}
}
\concept{metodos gamperda}
\concept{plots gamperda}
