% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgam.r
\name{optgam_prod}
\alias{optgam_prod}
\title{Otimização Do Ajuste De Produtibilidade}
\usage{
optgam_prod(
  dat,
  range = list(5:30, 5:30),
  ts = c("ps", "ps"),
  dist = gaussian(),
  bordas = TRUE,
  modo = c("tensor", "multivar", "simples")
)
}
\arguments{
\item{dat}{\code{data.table} de dados para ajuste. Ver Detalhes}

\item{range}{lista com dois vetores inteiros indicando dimensões de base em queda líquida e
vazão, nesta ordem, a testar na seleção. Todas as combinações serão testadas (um domínio
quadrado é gerado a partir dos ranges)}

\item{ts}{vetor de duas strings tipo de spline utilizada para queda líquida e vazão --
veja \code{\link[mgcv]{smooth.terms}} para mais detalhes sobre as opções}

\item{dist}{objeto da classe \code{family} indicando a distribuição e link a serem usados no
ajuste. Veja \code{\link[stats]{family}} e \code{\link[stats]{glm}} para mais informações.}

\item{bordas}{booleano indicando o uso ou não de bordas; alternativamente, um vetor de inteiros
indicando quais vértices utilizar. Ver Detalhes}

\item{modo}{um de \code{"tensor"}, \code{"multivar"} ou \code{"simples"} indicando o modo de
interação entre funções marginais. Ver Detalhes}
}
\value{
objeto \code{gamprod} contendo GAM e extrapolações estimadas
}
\description{
Varre uma faixa de dimensões de base, reportando aquela correspondente ao ajuste de menor BIC
}
\details{
Os parâmetros recebidos por \code{optgam_prod} são exatamente aqueles de
\code{\link{fitgam_prod}}, com uma exceção: o argumento \code{ns} é substituído por \code{range},
uma lista de dois vetores de inteiros indicando as dimensões de base candidatas em queda e vazão,
nesta ordem, para seleção. Mais detalhes a respeito da estimação e efeito dos argumentos
individuais podem ser encontrados em \code{\link{fitgam_prod}}.
}
\examples{

dat <- agregasemana(dummydata)

# execucao limitando a faixa de numero de splines a algo mais baixo
optfit <- optgam_prod(dat, range = list(5:10, 5:10))
\dontrun{
plot(optfit)
}

}
\seealso{
ajuste individual para uma dada dimensão de base \code{\link{fitgam_prod}}

Other metodos gamprod: 
\code{\link{fitgam_prod}()},
\code{\link{fitted.gamprod}()},
\code{\link{predict.gamprod}()},
\code{\link{residuals.gamprod}()}

Other plots gamprod: 
\code{\link{fitgam_prod}()},
\code{\link{plot.gamprod}()}
}
\concept{metodos gamprod}
\concept{plots gamprod}
