% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgrid.r
\name{optgrid}
\alias{optgrid}
\alias{optgrid.gamperda}
\title{Otimização De Segmentações De Grade}
\usage{
optgrid(fit, R = 1.01, ..., full.output = FALSE)

\method{optgrid}{gamperda}(fit, R = 1.01, range.vazao = 5:50, ..., full.output = FALSE)
}
\arguments{
\item{fit}{ajuste do dado de perda ou produtibilidade}

\item{R}{razão mínima entre erro do ajuste e da grade para aceitação}

\item{...}{demais parâmetros específicos de grade de perda ou produtibilidade. Ver métodos
Exemplos}

\item{full.output}{booleano, se verdadeiro retorna objeto de varredura}

\item{range.vazao}{vetor de inteiros indicando números de segmentações a serem testados}
}
\value{
se \code{full.output = FALSE}, a grade parsimoniosa que atenda a razão mínima \code{R}.
Se \code{full.output = TRUE}, uma lista com a grade no primeiro elemento e a varredura no
segundo
}
\description{
Otimiza o número de divisões de grade para atendimento de um critério mínimo
}
