% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optgrid.r
\name{optgrid}
\alias{optgrid}
\alias{optgrid.gamperda}
\alias{optgrid.gamprod}
\title{Otimização De Segmentações De Grade}
\usage{
optgrid(fit, R = 1.01, ..., full.output = FALSE)

\method{optgrid}{gamperda}(fit, R = 1.01, range.vazao = 5:50, ..., full.output = FALSE)

\method{optgrid}{gamprod}(
  fit,
  R = 1.01,
  range.quedal = 5:50,
  range.vazao = 5:50,
  ...,
  full.output = FALSE
)
}
\arguments{
\item{fit}{ajuste do dado de perda ou produtibilidade}

\item{R}{razão mínima entre erro do ajuste e da grade para aceitação}

\item{...}{demais parâmetros específicos de grade de perda ou produtibilidade. Ver métodos
Exemplos}

\item{full.output}{booleano, se verdadeiro retorna objeto de varredura}

\item{range.quedal, range.vazao}{vetores de inteiros indicando números de segmentações a serem
testadas em cada marginal. Todas as combinações serão testadas (isto é, um domínio quadrado é
gerado a partir dos ranges). Padrão \code{range.quedal = range.vazao = 5:50}}
}
\value{
se \code{full.output = FALSE}, a grade parsimoniosa que atenda a razão mínima \code{R}.
Se \code{full.output = TRUE}, uma lista com a grade no primeiro elemento e a varredura no
segundo
}
\description{
Otimiza o número de divisões de grade visando o atendimento de um critério mínimo
}
\details{
\code{optgrid} tenta determinar a melhor configuração de grade de modo que o erro de
representação do dado histórico por interpolação linear na grade seja suficientemente próximo do
erro do modelo latente ajustado. Isto se dá por uma pesquisa exaustiva, testando todas as
configurações de grade em uma determinada faixa de número de divisões permissíveis.

O argumento \code{R} controla o critério de comparação entre os erros de representação. Quando
\eqn{erro_grade / erro_modelo \le R} para um dado número de segmentações, considera-se esta
configuração adequada.

Para perdas, serão testados os números de divisões fornecidos pelo argumento \code{range.vazao},
um vetor de inteiros indicando quais números de segmentações testar. Será selecionado como número
ótimo o primeiro para o qual todos os números seguintes atendem ao critério de comparação. Isto é
feito para evitar convergências precipitadas, garantindo assim que o valor selecionado
corresponde à menor grade possível que atende confiavelmente ao critério de comparação.
}
\examples{

# otimizando grade de perdas -----------------------------------------------

dat <- agregasemana(dummydata)
ajuste <- fitgam_perda(dat, ns.vazao = 5)

# roda otimizacao do numero de segmentacoes sem retornar varredura
optgrade <- optgrid(ajuste, range.vazao = 10:60)

# assim como ajustes de curvas, grades tambem possuem um metodo de plot
\dontrun{
plot(optgrade)
}

# e possivel rodar a otimizacao retornando a varredura de numero de segmentacoes e plota-la
optgrade <- optgrid(ajuste, range.vazao = 10:60, full.output = TRUE)
\dontrun{
plot(optgrade[[2]])
}

}
\seealso{
extração de grade para dadas dimensões \code{\link{extraigrid}}. Métodos para
plot da varredura quando \code{full.output = TRUE}: \code{\link{plot.varreduraperda}} e
\code{\link{plot.varreduraprod}}

Other metodos gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{fitted.gridperda}()},
\code{\link{predict.gridperda}()},
\code{\link{residuals.gridperda}()}

Other plots gridperda: 
\code{\link{extraigrid.gamperda}()},
\code{\link{plot.gridperda}()}

Other metodos gridpprod: 
\code{\link{extraigrid.gamperda}()}

Other plots gridpprod: 
\code{\link{extraigrid.gamperda}()}
}
\concept{metodos gridperda}
\concept{metodos gridpprod}
\concept{plots gridperda}
\concept{plots gridpprod}
