% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizacao.r
\name{plot.gamperda}
\alias{plot.gamperda}
\title{Plot Completo De Objetos \code{gamperda}}
\usage{
\method{plot}{gamperda}(x, legenda = TRUE, ...)
}
\arguments{
\item{x}{objeto da classe \code{gamperda}}

\item{legenda}{booleano indicando se uma legenda automática deve ser adicionada}

\item{...}{existe apenas para consistência com a genérica}
}
\value{
plota dados originais e ajuste realizado
}
\description{
Wrapper para visualização dos modelos contínuos de perda junto aos dados ajustados
}
\examples{

dat <- agregasemana(dummydata)
mod <- fitgam_perda(dat)

\dontrun{
plot(mod)
}

}
\seealso{
Other plots gamperda: 
\code{\link{fitgam_perda}()},
\code{\link{lines.gamperda}()},
\code{\link{optgam_perda}()}
}
\concept{plots gamperda}
