% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gamperda.r
\name{print.gamperda}
\alias{print.gamperda}
\title{Print De Objetos \code{gamperda}}
\usage{
\method{print}{gamperda}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{gamperda}}
}
\description{
Print De Objetos \code{gamperda}
}
