% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classe-gridperda.r
\name{print.gridperda}
\alias{print.gridperda}
\title{Print De Objetos \code{gridperda}}
\usage{
\method{print}{gridperda}(x, ...)
}
\arguments{
\item{x}{objeto da classe \code{gridperda}}
}
\description{
Print De Objetos \code{gridperda}
}
